Move Sector to its own class
