+++ PyRoute/DeltaDebug/DeltaGalaxy.py
@@ -6,8 +6,9 @@ Created on May 21, 2023
 
 from PyRoute.AreaItems.Allegiance import Allegiance
 from PyRoute.AreaItems.Subsector import Subsector
+from PyRoute.AreaItems.Sector import Sector
 from PyRoute.DeltaDebug.DeltaDictionary import SectorDictionary, SubsectorDictionary
-from PyRoute.Galaxy import Galaxy, Sector
+from PyRoute.Galaxy import Galaxy
 from PyRoute.Star import Star
 from PyRoute.AllyGen import AllyGen
