+++ PyRoute/Galaxy.py
@@ -3,7 +3,6 @@ Created on Mar 2, 2014
 
 @author: tjoneslo
 """
-import copy
 import logging
 import re
 import codecs
@@ -18,6 +17,7 @@ from PyRoute.AllyGen import AllyGen
 from PyRoute.AreaItems.AreaItem import AreaItem
 from PyRoute.AreaItems.Allegiance import Allegiance
 from PyRoute.AreaItems.Subsector import Subsector
+from PyRoute.AreaItems.Sector import Sector
 from PyRoute.Position.Hex import Hex
 from PyRoute.Star import Star
 from PyRoute.Calculation.TradeCalculation import TradeCalculation
@@ -28,129 +28,6 @@ from PyRoute.Calculation.NoneCalculation import NoneCalculation
 from PyRoute.Calculation.XRouteCalculation import XRouteCalculation
 from PyRoute.DataClasses.ReadSectorOptions import ReadSectorOptions
 from PyRoute.Pathfinding.RouteLandmarkGraph import RouteLandmarkGraph
-
-
-class Sector(AreaItem):
-    position_match = re.compile(r"([-+ ]?\d{1,4})\s*[,]\s*([-+ ]?\d{1,4})")
-
-    def __init__(self, name, position):
-        name = name.rstrip()
-        if 3 > len(name):
-            raise ValueError("Name string too short")
-        if 5 > len(position):
-            raise ValueError("Position string too short")
-        if '#' != name[0]:
-            raise ValueError("Name string should start with #")
-        if '#' != position[0]:
-            raise ValueError("Position string should start with #")
-
-        # Pre-trim input values
-
-        # The name as passed from the Galaxy read include the comment marker at the start of the line
-        # So strip the comment marker, then strip spaces.
-        super(Sector, self).__init__(name[1:].strip())
-
-        # Same here, the position has a leading comment marker.  Strip that, and then any flanking whitespace.
-        position = position[1:].strip().replace(' ', '')
-        positions = Sector.position_match.match(position)
-        if not positions:
-            raise ValueError("Position string malformed")
-
-        pos_bits = positions.groups()
-
-        self._wiki_name = '[[{0} Sector|{0}]]'.format(self.sector_name())
-
-        self.x = int(pos_bits[0])
-        self.y = int(pos_bits[1])
-
-        self.dx = self.x * 32
-        self.dy = self.y * 40
-        self.subsectors = {}
-        self.spinward = None
-        self.trailing = None
-        self.coreward = None
-        self.rimward = None
-
-    # For the JSONPickel work
-    def __getstate__(self):
-        state = self.__dict__.copy()
-        del state['spinward']
-        del state['trailing']
-        del state['coreward']
-        del state['rimward']
-        del state['alg_sorted']
-        return state
-
-    def __str__(self):
-        return '{} ({},{})'.format(self.name, str(self.x), str(self.y))
-
-    def sector_name(self):
-        return self.name[:-7] if self.name.endswith('Sector') else self.name
-
-    def find_world_by_pos(self, pos):
-        for world in self.worlds:
-            if world.position == pos:
-                return world
-        return None
-
-    def is_well_formed(self):
-        msg = ""
-        # check name
-        if self.name is None or '' == self.name.strip():
-            msg = "Name cannot be empty"
-            return False, msg
-
-        # Check reciprocity of adjacent sectors
-        if self.coreward is not None:
-            neighbour = self.coreward
-            other = neighbour.rimward
-            if other is None or other != self:
-                msg = "Coreward sector mismatch for " + self.name
-                return False, msg
-            if self.x != neighbour.x:
-                msg = "Coreward sector x co-ord mismatch for " + self.name
-                return False, msg
-            if self.y + 1 != neighbour.y:
-                msg = "Coreward sector y co-ord mismatch for " + self.name
-                return False, msg
-        if self.rimward is not None:
-            neighbour = self.rimward
-            other = neighbour.coreward
-            if other is None or other != self:
-                msg = "Rimward sector mismatch for " + self.name
-                return False, msg
-            if self.x != neighbour.x:
-                msg = "Rimward sector x co-ord mismatch for " + self.name
-                return False, msg
-            if self.y - 1 != neighbour.y:
-                msg = "Rimward sector y co-ord mismatch for " + self.name
-                return False, msg
-        if self.spinward is not None:
-            neighbour = self.spinward
-            other = neighbour.trailing
-            if other is None or other != self:
-                msg = "Spinward sector mismatch for " + self.name
-                return False, msg
-            if self.x - 1 != neighbour.x:
-                msg = "Spinward sector x co-ord mismatch for " + self.name
-                return False, msg
-            if self.y != neighbour.y:
-                msg = "Spinward sector y co-ord mismatch for " + self.name
-                return False, msg
-        if self.trailing is not None:
-            neighbour = self.trailing
-            other = neighbour.spinward
-            if other is None or other != self:
-                msg = "Trailing sector mismatch for " + self.name
-                return False, msg
-            if self.x + 1 != neighbour.x:
-                msg = "Trailing sector x co-ord mismatch for " + self.name
-                return False, msg
-            if self.y != neighbour.y:
-                msg = "Trailing sector y co-ord mismatch for " + self.name
-                return False, msg
-
-        return True, msg
 
 
 class Galaxy(AreaItem):
