+++ Tests/Hypothesis/testTradeCodes.py
@@ -4,8 +4,7 @@ from datetime import timedelta
 from hypothesis import given, assume, example, HealthCheck, settings
 from hypothesis.strategies import text, from_regex, composite, sampled_from, lists, floats, booleans
 
-from PyRoute.Galaxy import Sector
-from PyRoute.Inputs.ParseStarInput import ParseStarInput
+from PyRoute.AreaItems.Sector import Sector
 from PyRoute.TradeCodes import TradeCodes
 from PyRoute.Star import Star
 from PyRoute.SystemData.UWP import UWP
