+++ Tests/testStarDistances.py
@@ -1,7 +1,7 @@
 import unittest
 
+from PyRoute.AreaItems.Sector import Sector
 from PyRoute.Position.Hex import Hex
-from PyRoute.Galaxy import Sector
 from PyRoute.Star import Star
 
 distance_list = [
